scripts/generate_script.py
#!/usr/bin/env python3
"""
スライドから原稿を生成するスクリプト
Gemini APIを使用してスライドの内容から自然な原稿を生成します

YAMLにscript_notesが含まれている場合はそれを直接使用し、
含まれていない場合はAIで原稿を生成します。
"""

import sys
import os
import json
import re
import time
import yaml
import argparse
from pathlib import Path
import google.generativeai as genai
from google.api_core import exceptions as google_exceptions


def load_script_notes_from_yaml(yaml_file):
    """
    YAMLファイルからscript_notesを読み込む

    Args:
        yaml_file: YAMLファイルのパス

    Returns:
        script_notesのリスト、またはNone（script_notesがない場合）
    """
    try:
        with open(yaml_file, 'r', encoding='utf-8') as f:
            data = yaml.safe_load(f)

        if 'script_notes' in data and data['script_notes']:
            slides = data.get('slides', [])
            script_notes = data['script_notes']

            # script_notesとslidesを組み合わせる
            scripts = []
            for note in script_notes:
                slide_num = note.get('slide', len(scripts) + 1)
                # 対応するスライドのタイトルを取得
                title = ''
                if slide_num <= len(slides):
                    title = slides[slide_num - 1].get('title', f'スライド {slide_num}')

                scripts.append({
                    'index': slide_num,
                    'title': title,
                    'script': note.get('script', '')
                })

            return scripts
    except Exception as e:
        print(f"YAMLからのscript_notes読み込みに失敗: {e}")

    return None

def parse_marp_slides(slide_file):
    """
    Marp形式のMarkdownスライドを解析してスライドごとに分割

    Args:
        slide_file: スライドファイルのパス

    Returns:
        スライドのリスト（各スライドは辞書形式）
    """
    with open(slide_file, 'r', encoding='utf-8') as f:
        content = f.read()

    # スライドを分割（---で区切られている）
    slides_raw = content.split('---')

    slides = []
    for i, slide_raw in enumerate(slides_raw):
        slide_raw = slide_raw.strip()
        if not slide_raw or slide_raw.startswith('marp:'):
            continue

        # タイトルと内容を抽出
        lines = slide_raw.split('\n')
        title = ''
        content_lines = []

        for line in lines:
            if line.startswith('# '):
                title = line[2:].strip()
            elif line.strip():
                # <style>タグ、画像タグ、空行は内容から除外
                if not (line.strip().startswith('<style>') or
                        line.strip().startswith('</style>') or
                        line.strip().startswith('![') or
                        line.strip().startswith('@import') or
                        line.strip().startswith('section {') or
                        line.strip().startswith('h1,') or
                        line.strip().startswith('font-family:') or
                        line.strip() == '}'):
                    content_lines.append(line)

        # <style>タグのみのセクションはスキップ
        content_text = '\n'.join(content_lines).strip()
        if not title and not content_text:
            continue

        slides.append({
            'index': len(slides) + 1,
            'title': title,
            'content': content_text
        })

    return slides

def clean_declarative_phrases(script):
    """
    原稿から不要な宣言的フレーズを削除

    Args:
        script: 生成された原稿テキスト

    Returns:
        クリーンアップされた原稿テキスト
    """
    # 削除するパターン（正規表現）
    patterns_to_remove = [
        # スライド作成・番号に関する言及
        r'^.*?原稿を作成.*?[。\n]',
        r'^.*?スライド\d+.*?[。\n]',

        # 接続詞での始まり（スライド遷移の言及）
        r'^(では|それでは|次に|続いて|さて|それから)[、，]?.*?[。\n]',

        # プレゼン構造の言及
        r'^.*?(このスライド|今回|今日|本日|ここ)(では|で|から|について).*?[。\n]',

        # その他のメタ的な言及
        r'^.*?(ご紹介|説明|見て|ご覧|解説)(します|いたします|しましょう|ください).*?[。\n]',
    ]

    cleaned = script
    for pattern in patterns_to_remove:
        cleaned = re.sub(pattern, '', cleaned, flags=re.MULTILINE | re.IGNORECASE)

    # 先頭の空白行を削除
    cleaned = cleaned.lstrip('\n').strip()

    return cleaned

def generate_script_for_slide(model, slide, total_slides, max_retries=3):
    """
    1枚のスライドに対する原稿を生成（リトライ機能付き）

    Args:
        model: Gemini モデル
        slide: スライド情報（辞書）
        total_slides: 総スライド数
        max_retries: 最大リトライ回数

    Returns:
        生成された原稿テキスト
    """
    prompt = f"""
あなたはプレゼンテーションの原稿を作成する専門家です。
以下のスライド情報から、自然で分かりやすい原稿を日本語で作成してください。

スライド番号: {slide['index']} / {total_slides}
タイトル: {slide['title']}
内容:
{slide['content']}

要件:
1. 自然な話し言葉で書いてください
2. 1スライドあたり30〜60秒程度の長さにしてください
3. 箇条書きは自然な文章に変換してください
4. 聞き手に語りかけるような口調にしてください
5. 「えー」「あのー」などの言葉は入れないでください
6. スライド番号や「このスライドでは」などの言及は避けてください

原稿のみを出力してください（説明や補足は不要です）。
"""

    for attempt in range(max_retries):
        try:
            response = model.generate_content(prompt)
            script = response.text.strip()

            # 不要な宣言的フレーズを削除
            script = clean_declarative_phrases(script)

            return script
        except google_exceptions.ResourceExhausted as e:
            if attempt < max_retries - 1:
                wait_time = 45  # レート制限エラーの場合は45秒待機
                print(f"  レート制限エラー（試行 {attempt + 1}/{max_retries}）")
                print(f"  {wait_time}秒待機してリトライします...")
                time.sleep(wait_time)
            else:
                print(f"  最大リトライ回数に達しました。エラー: {e}")
                raise
        except Exception as e:
            if attempt < max_retries - 1:
                wait_time = 10
                print(f"  エラー発生（試行 {attempt + 1}/{max_retries}）: {str(e)[:100]}")
                print(f"  {wait_time}秒待機してリトライします...")
                time.sleep(wait_time)
            else:
                print(f"  最大リトライ回数に達しました。エラー: {e}")
                raise

def generate_full_script(slide_file, output_file, yaml_file=None):
    """
    スライドファイル全体の原稿を生成

    Args:
        slide_file: スライドファイルのパス
        output_file: 出力ファイルのパス
        yaml_file: YAMLファイルのパス（script_notesがある場合に使用）
    """
    scripts = None

    # YAMLからscript_notesを読み込む（利用可能な場合）
    if yaml_file and os.path.exists(yaml_file):
        print(f"YAMLファイルからscript_notesを読み込み中: {yaml_file}")
        scripts = load_script_notes_from_yaml(yaml_file)
        if scripts:
            print(f"✓ script_notesを使用します（{len(scripts)}スライド分）")
        else:
            print("YAMLにscript_notesがないため、AIで生成します")

    # script_notesがない場合はAIで生成
    if not scripts:
        # APIキーの確認
        api_key = os.environ.get('GEMINI_API_KEY')
        if not api_key:
            raise ValueError("GEMINI_API_KEY環境変数が設定されていません")

        # Gemini APIの初期化
        genai.configure(api_key=api_key)
        model = genai.GenerativeModel('gemini-2.0-flash-exp')

        # スライドを解析
        print(f"スライドを解析中: {slide_file}")
        slides = parse_marp_slides(slide_file)
        print(f"スライド数: {len(slides)}")

        # 各スライドの原稿を生成
        scripts = []
        for i, slide in enumerate(slides):
            print(f"原稿生成中: スライド {slide['index']} - {slide['title']}")
            script = generate_script_for_slide(model, slide, len(slides))

            scripts.append({
                'index': slide['index'],
                'title': slide['title'],
                'script': script
            })

            print(f"  生成完了: {len(script)}文字")

            # レート制限対策：各スライド間に7秒待機（最後のスライドを除く）
            # Gemini APIは1分間に10リクエストまでなので、6秒に1リクエストのペース
            if i < len(slides) - 1:
                wait_time = 7
                print(f"  レート制限対策のため{wait_time}秒待機中...")
                time.sleep(wait_time)

    # JSONとして保存
    output_data = {
        'slides': scripts,
        'total_slides': len(scripts)
    }

    with open(output_file, 'w', encoding='utf-8') as f:
        json.dump(output_data, f, ensure_ascii=False, indent=2)

    print(f"\n原稿を保存しました: {output_file}")

    # テキスト版も保存
    text_output = str(output_file).replace('.json', '.txt')
    with open(text_output, 'w', encoding='utf-8') as f:
        for script in scripts:
            f.write(f"=== スライド {script['index']}: {script['title']} ===\n\n")
            f.write(script['script'])
            f.write("\n\n")

    print(f"テキスト版も保存しました: {text_output}")

    return output_file

def main():
    parser = argparse.ArgumentParser(
        description='スライドから原稿を生成するスクリプト'
    )
    parser.add_argument('slide_file', help='スライドファイルのパス')
    parser.add_argument(
        '--yaml', '-y',
        help='YAMLファイルのパス（script_notesがある場合に使用）',
        default=None
    )

    args = parser.parse_args()

    slide_file = args.slide_file
    yaml_file = args.yaml

    if not os.path.exists(slide_file):
        print(f"エラー: スライドファイルが見つかりません: {slide_file}")
        sys.exit(1)

    # 出力ディレクトリとファイル名
    slide_path = Path(slide_file)
    output_dir = slide_path.parent.parent / "scripts_output"
    output_dir.mkdir(exist_ok=True)

    output_file = output_dir / f"{slide_path.stem}_script.json"

    # 原稿を生成
    script_file = generate_full_script(slide_file, output_file, yaml_file)

    # GitHub Actions用に環境変数に保存
    if 'GITHUB_ENV' in os.environ:
        with open(os.environ['GITHUB_ENV'], 'a') as f:
            f.write(f"SCRIPT_FILE={script_file}\n")

if __name__ == "__main__":
    main()
