requirements.txt
requests>=2.31.0
feedparser>=6.0.10
google-generativeai>=0.3.0
google-genai
google-api-python-client>=2.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
pandas>=2.1.0
markdown>=3.5.0
weasyprint>=60.0
pygments>=2.17.0
yfinance>=0.2.0
pyyaml
gTTS
pydub
pillow
edge-tts
